Report integration test

+++ src/main/java/com/gowpet/pos/billing/service/AggregatedBillingItem.java
@@ -1,4 +1,4 @@
 package com.gowpet.pos.billing.service;
 
-public record AggregatedBillingItem (String catalogItemId, Double price, Long quantity) {
+public record AggregatedBillingItem (String catalogItemId, Double price, Double quantity) {
 }
